DASH-853 Move snap action forms to components folder.
