+++ client/src/landing-pages/fio-wallet-snap-landing-page/FioWalletSnapComponent/components/FioActionForms/CustomActionForm.tsx
@@ -1,34 +1,26 @@
 import React from 'react';
-
 import { Form, Field } from 'react-final-form';
 
 import TextInput, {
   INPUT_COLOR_SCHEMA,
   INPUT_UI_STYLES,
-} from '../../../../components/Input/TextInput';
-import { COLOR_TYPE } from '../../../../components/Input/ErrorBadge';
-import SubmitButton from '../../../../components/common/SubmitButton/SubmitButton';
+} from '../../../../../components/Input/TextInput';
+import { COLOR_TYPE } from '../../../../../components/Input/ErrorBadge';
+import SubmitButton from '../../../../../components/common/SubmitButton/SubmitButton';
 
 type ErrorsProps = {
-  amount?: string;
-  fioHandle?: string;
+  customAction?: string;
 };
 
 const validateForm = ({
-  amount,
-  fioHandle,
+  customAction,
 }: {
-  amount: string;
-  fioHandle: string;
+  customAction: string;
 }): ErrorsProps => {
   const errors: ErrorsProps = {};
 
-  if (!amount) {
-    errors.amount = 'Required';
-  }
-
-  if (!fioHandle) {
-    errors.fioHandle = 'Required';
+  if (!customAction) {
+    errors.customAction = 'Required';
   }
 
   return errors;
@@ -38,7 +30,7 @@ type Props = {
   onSubmit: (values: any) => void;
 };
 
-export const StakeFioTokens: React.FC<Props> = props => {
+export const CustomActionForm: React.FC<Props> = props => {
   const { onSubmit } = props;
   return (
     <div>
@@ -49,21 +41,12 @@ export const StakeFioTokens: React.FC<Props> = props => {
           return (
             <form onSubmit={formProps.handleSubmit}>
               <Field
-                name="amount"
-                type="text"
+                name="customAction"
+                type="textarea"
+                element="textarea"
                 component={TextInput}
-                placeholder="Type amount to stake"
-                label="FIO Tokens amount"
-                colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
-                uiType={INPUT_UI_STYLES.INDIGO_WHITE}
-                errorColor={COLOR_TYPE.WARN}
-              />
-              <Field
-                name="fioHandle"
-                type="text"
-                component={TextInput}
-                placeholder="Type FIO Handle"
-                label="FIO Handle"
+                placeholder="Type Custom Action Params"
+                label="Custom Action Params"
                 colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
                 uiType={INPUT_UI_STYLES.INDIGO_WHITE}
                 errorColor={COLOR_TYPE.WARN}
