+++ client/src/landing-pages/fio-wallet-snap-landing-page/FioWalletSnapComponent/components/FioActionForms/RegisterFioDomain.tsx
@@ -1,27 +1,22 @@
 import React from 'react';
-
 import { Form, Field } from 'react-final-form';
 
 import TextInput, {
   INPUT_COLOR_SCHEMA,
   INPUT_UI_STYLES,
-} from '../../../../components/Input/TextInput';
-import { COLOR_TYPE } from '../../../../components/Input/ErrorBadge';
-import SubmitButton from '../../../../components/common/SubmitButton/SubmitButton';
+} from '../../../../../components/Input/TextInput';
+import { COLOR_TYPE } from '../../../../../components/Input/ErrorBadge';
+import SubmitButton from '../../../../../components/common/SubmitButton/SubmitButton';
 
 type ErrorsProps = {
-  fioRequestId?: string;
+  fioDomain?: string;
 };
 
-const validateForm = ({
-  fioRequestId,
-}: {
-  fioRequestId: string;
-}): ErrorsProps => {
+const validateForm = ({ fioDomain }: { fioDomain: string }): ErrorsProps => {
   const errors: ErrorsProps = {};
 
-  if (!fioRequestId) {
-    errors.fioRequestId = 'Required';
+  if (!fioDomain) {
+    errors.fioDomain = 'Required';
   }
 
   return errors;
@@ -31,7 +26,7 @@ type Props = {
   onSubmit: (values: any) => void;
 };
 
-export const RejectFioRequest: React.FC<Props> = props => {
+export const RegisterFioDomain: React.FC<Props> = props => {
   const { onSubmit } = props;
   return (
     <div>
@@ -42,11 +37,11 @@ export const RejectFioRequest: React.FC<Props> = props => {
           return (
             <form onSubmit={formProps.handleSubmit}>
               <Field
-                name="fioRequestId"
+                name="fioDomain"
                 type="text"
                 component={TextInput}
-                placeholder="Type FIO Request id"
-                label="FIO Request Id"
+                placeholder="Type FIO Domain"
+                label="FIO Domain"
                 colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
                 uiType={INPUT_UI_STYLES.INDIGO_WHITE}
                 errorColor={COLOR_TYPE.WARN}
