+++ client/src/landing-pages/fio-wallet-snap-landing-page/FioWalletSnapComponent/components/FioActionForms/RegisterFioHandle.tsx
@@ -4,19 +4,19 @@ import { Form, Field } from 'react-final-form';
 import TextInput, {
   INPUT_COLOR_SCHEMA,
   INPUT_UI_STYLES,
-} from '../../../../components/Input/TextInput';
-import { COLOR_TYPE } from '../../../../components/Input/ErrorBadge';
-import SubmitButton from '../../../../components/common/SubmitButton/SubmitButton';
+} from '../../../../../components/Input/TextInput';
+import { COLOR_TYPE } from '../../../../../components/Input/ErrorBadge';
+import SubmitButton from '../../../../../components/common/SubmitButton/SubmitButton';
 
 type ErrorsProps = {
-  fioDomain?: string;
+  fioHandle?: string;
 };
 
-const validateForm = ({ fioDomain }: { fioDomain: string }): ErrorsProps => {
+const validateForm = ({ fioHandle }: { fioHandle: string }): ErrorsProps => {
   const errors: ErrorsProps = {};
 
-  if (!fioDomain) {
-    errors.fioDomain = 'Required';
+  if (!fioHandle) {
+    errors.fioHandle = 'Required';
   }
 
   return errors;
@@ -26,7 +26,7 @@ type Props = {
   onSubmit: (values: any) => void;
 };
 
-export const RenewFioDomain: React.FC<Props> = props => {
+export const RegisterFioHandle: React.FC<Props> = props => {
   const { onSubmit } = props;
   return (
     <div>
@@ -37,11 +37,11 @@ export const RenewFioDomain: React.FC<Props> = props => {
           return (
             <form onSubmit={formProps.handleSubmit}>
               <Field
-                name="fioDomain"
+                name="fioHandle"
                 type="text"
                 component={TextInput}
-                placeholder="Type FIO Domain"
-                label="FIO Domain"
+                placeholder="Type FIO Handle"
+                label="FIO Handle"
                 colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
                 uiType={INPUT_UI_STYLES.INDIGO_WHITE}
                 errorColor={COLOR_TYPE.WARN}
