+++ client/src/landing-pages/fio-wallet-snap-landing-page/FioWalletSnapComponent/components/FioActionForms/TransferFioDomain.tsx
@@ -4,19 +4,30 @@ import { Form, Field } from 'react-final-form';
 import TextInput, {
   INPUT_COLOR_SCHEMA,
   INPUT_UI_STYLES,
-} from '../../../../components/Input/TextInput';
-import { COLOR_TYPE } from '../../../../components/Input/ErrorBadge';
-import SubmitButton from '../../../../components/common/SubmitButton/SubmitButton';
+} from '../../../../../components/Input/TextInput';
+import { COLOR_TYPE } from '../../../../../components/Input/ErrorBadge';
+import SubmitButton from '../../../../../components/common/SubmitButton/SubmitButton';
 
 type ErrorsProps = {
   fioDomain?: string;
+  newOwnerPublicKey?: string;
 };
 
-const validateForm = ({ fioDomain }: { fioDomain: string }): ErrorsProps => {
+const validateForm = ({
+  fioDomain,
+  newOwnerPublicKey,
+}: {
+  fioDomain: string;
+  newOwnerPublicKey: string;
+}): ErrorsProps => {
   const errors: ErrorsProps = {};
 
   if (!fioDomain) {
     errors.fioDomain = 'Required';
+  }
+
+  if (!newOwnerPublicKey) {
+    errors.newOwnerPublicKey = 'Required';
   }
 
   return errors;
@@ -26,7 +37,7 @@ type Props = {
   onSubmit: (values: any) => void;
 };
 
-export const RegisterFioDomain: React.FC<Props> = props => {
+export const TransferFioDomain: React.FC<Props> = props => {
   const { onSubmit } = props;
   return (
     <div>
@@ -42,6 +53,16 @@ export const RegisterFioDomain: React.FC<Props> = props => {
                 component={TextInput}
                 placeholder="Type FIO Domain"
                 label="FIO Domain"
+                colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
+                uiType={INPUT_UI_STYLES.INDIGO_WHITE}
+                errorColor={COLOR_TYPE.WARN}
+              />
+              <Field
+                name="newOwnerPublicKey"
+                type="text"
+                component={TextInput}
+                placeholder="Type New Owner Public Key"
+                label="FIO Public Key"
                 colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
                 uiType={INPUT_UI_STYLES.INDIGO_WHITE}
                 errorColor={COLOR_TYPE.WARN}
