+++ client/src/landing-pages/fio-wallet-snap-landing-page/FioWalletSnapComponent/components/FioActionForms/UnstakeFioTokens.tsx
@@ -5,24 +5,21 @@ import { Form, Field } from 'react-final-form';
 import TextInput, {
   INPUT_COLOR_SCHEMA,
   INPUT_UI_STYLES,
-} from '../../../../components/Input/TextInput';
-import { COLOR_TYPE } from '../../../../components/Input/ErrorBadge';
-import SubmitButton from '../../../../components/common/SubmitButton/SubmitButton';
+} from '../../../../../components/Input/TextInput';
+import { COLOR_TYPE } from '../../../../../components/Input/ErrorBadge';
+import SubmitButton from '../../../../../components/common/SubmitButton/SubmitButton';
 
 type ErrorsProps = {
   amount?: string;
-  payeeFioHandle?: string;
-  payerFioHandle?: string;
+  fioHandle?: string;
 };
 
 const validateForm = ({
   amount,
-  payeeFioHandle,
-  payerFioHandle,
+  fioHandle,
 }: {
   amount: string;
-  payeeFioHandle: string;
-  payerFioHandle: string;
+  fioHandle: string;
 }): ErrorsProps => {
   const errors: ErrorsProps = {};
 
@@ -30,12 +27,8 @@ const validateForm = ({
     errors.amount = 'Required';
   }
 
-  if (!payeeFioHandle) {
-    errors.payeeFioHandle = 'Required';
-  }
-
-  if (!payerFioHandle) {
-    errors.payerFioHandle = 'Required';
+  if (!fioHandle) {
+    errors.fioHandle = 'Required';
   }
 
   return errors;
@@ -45,7 +38,7 @@ type Props = {
   onSubmit: (values: any) => void;
 };
 
-export const RequestNewFunds: React.FC<Props> = props => {
+export const UnstakeFioTokens: React.FC<Props> = props => {
   const { onSubmit } = props;
   return (
     <div>
@@ -56,41 +49,21 @@ export const RequestNewFunds: React.FC<Props> = props => {
           return (
             <form onSubmit={formProps.handleSubmit}>
               <Field
-                name="payeeFioHandle"
-                type="text"
-                component={TextInput}
-                placeholder="Type payee FIO Handle"
-                label="Payee FIO Handle"
-                colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
-                uiType={INPUT_UI_STYLES.INDIGO_WHITE}
-                errorColor={COLOR_TYPE.WARN}
-              />
-              <Field
-                name="payerFioHandle"
-                type="text"
-                component={TextInput}
-                placeholder="Type payer FIO Handle"
-                label="Payer FIO Handle"
-                colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
-                uiType={INPUT_UI_STYLES.INDIGO_WHITE}
-                errorColor={COLOR_TYPE.WARN}
-              />
-              <Field
                 name="amount"
                 type="text"
                 component={TextInput}
-                placeholder="Type requested amount"
-                label="FIO tokens amount"
+                placeholder="Type amount to unstake"
+                label="FIO Tokens amount"
                 colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
                 uiType={INPUT_UI_STYLES.INDIGO_WHITE}
                 errorColor={COLOR_TYPE.WARN}
               />
               <Field
-                name="memo"
+                name="fioHandle"
                 type="text"
                 component={TextInput}
-                placeholder="Type memo"
-                label="Memo"
+                placeholder="Type FIO Handle"
+                label="FIO Handle"
                 colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
                 uiType={INPUT_UI_STYLES.INDIGO_WHITE}
                 errorColor={COLOR_TYPE.WARN}
