+++ client/src/landing-pages/fio-wallet-snap-landing-page/FioWalletSnapComponent/components/FioActionForms/WrapFioTokens.tsx
@@ -1,25 +1,27 @@
 import React from 'react';
-
 import { Form, Field } from 'react-final-form';
 
 import TextInput, {
   INPUT_COLOR_SCHEMA,
   INPUT_UI_STYLES,
-} from '../../../../components/Input/TextInput';
-import { COLOR_TYPE } from '../../../../components/Input/ErrorBadge';
-import SubmitButton from '../../../../components/common/SubmitButton/SubmitButton';
+} from '../../../../../components/Input/TextInput';
+import { COLOR_TYPE } from '../../../../../components/Input/ErrorBadge';
+import SubmitButton from '../../../../../components/common/SubmitButton/SubmitButton';
 
 type ErrorsProps = {
   amount?: string;
-  fioHandle?: string;
+  chainCode?: string;
+  publicAddress?: string;
 };
 
 const validateForm = ({
   amount,
-  fioHandle,
+  chainCode,
+  publicAddress,
 }: {
   amount: string;
-  fioHandle: string;
+  chainCode: string;
+  publicAddress: string;
 }): ErrorsProps => {
   const errors: ErrorsProps = {};
 
@@ -27,8 +29,12 @@ const validateForm = ({
     errors.amount = 'Required';
   }
 
-  if (!fioHandle) {
-    errors.fioHandle = 'Required';
+  if (!chainCode) {
+    errors.chainCode = 'Required';
+  }
+
+  if (!publicAddress) {
+    errors.publicAddress = 'Required';
   }
 
   return errors;
@@ -38,7 +44,7 @@ type Props = {
   onSubmit: (values: any) => void;
 };
 
-export const UnstakeFioTokens: React.FC<Props> = props => {
+export const WrapFioTokens: React.FC<Props> = props => {
   const { onSubmit } = props;
   return (
     <div>
@@ -52,18 +58,28 @@ export const UnstakeFioTokens: React.FC<Props> = props => {
                 name="amount"
                 type="text"
                 component={TextInput}
-                placeholder="Type amount to unstake"
-                label="FIO Tokens amount"
+                placeholder="Type FIO Tokens Amount"
+                label="FIO Tokens Amount"
                 colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
                 uiType={INPUT_UI_STYLES.INDIGO_WHITE}
                 errorColor={COLOR_TYPE.WARN}
               />
               <Field
-                name="fioHandle"
+                name="chainCode"
                 type="text"
                 component={TextInput}
-                placeholder="Type FIO Handle"
-                label="FIO Handle"
+                placeholder="Type Chain Code where to wrap"
+                label="Chain Code"
+                colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
+                uiType={INPUT_UI_STYLES.INDIGO_WHITE}
+                errorColor={COLOR_TYPE.WARN}
+              />
+              <Field
+                name="publicAddress"
+                type="text"
+                component={TextInput}
+                placeholder="Type Public Address"
+                label="Public Address"
                 colorSchema={INPUT_COLOR_SCHEMA.INDIGO_AND_WHITE}
                 uiType={INPUT_UI_STYLES.INDIGO_WHITE}
                 errorColor={COLOR_TYPE.WARN}
